requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-core
python-dotenv
python-docx
